backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
dedalus-labs
httpx==0.27.0
motor==3.7.1
PyJWT==2.10.1
bcrypt==5.0.0

